Removed caching from the declaredPropertiesForSubclass method.
I don’t know why I thought caching was needed on this to begin with.

Objective-C’s dynamic nature makes it so that a property can be added
at any time so caching a list of a class’ properties would only work if
I knew when to invalidate that list.

In addition, caching in general may need to be removed for all
instances of FDDeclaredProperty because I am not 100% sure that a
property’s attribute string can not be changed at runtime.